+++ src/App.vue
@@ -1,14 +1,17 @@
 <script setup lang="js">
 import {generateLargeComplexObject} from '../script/generateData.js';
 import _ from 'lodash';
-import { dataFind } from '../lib';
-const data = generateLargeComplexObject();
-const arr = [1,2,3,4,5,6,7,8,9,10];
-const newArr = dataFind({ a: 1, b: { c: null , d: 3} }, 'c', null); // { c: 2, d: 3 }
-console.log(newArr);
+import { getStorage,getCookie, setCookie, removeCookie } from '../lib';
+
+  setCookie('test',{name:'yi'},{
+    expires: 60 * 1000,
+  })
+    console.log(getCookie('test'))
+
 </script>
 
 <template>
   <div>
+    <button @click="copyText">button</button>
   </div>
 </template>